mcp-servers/linkedin-mcp-server/pyproject.toml
[project]
name = "linkedin-mcp-server"
version = "0.1.0"
description = "MCP server for posting to LinkedIn"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
linkedin-mcp-server = "linkedin_mcp_server.server:main"
